Lua: 'strcoll' -> 'strcmp' replacement to get rid of locale dependencies
This change is needed to prevent locale specific differences
in string comparison. This was causing some problems due to
differences in crc32 hashes calculation for maps and weapon
sets in ufo2000 on different machines (crc32 calculation for
the content of lua table uses string sorting of table keys).

An alternative solution would be to use a custom sorting function
for table keys when calculating crc32 for lua tables. But it is
better to prevent this source of inconsistencies and ensure
identical lua scripts execution for this type of operation on
different machines. Also it is more consistent with the way how
std::string comparison works in C++ (a primary programming
language for ufo2000).

From: Siarhei Siamashka <siarhei.siamashka@gmail.com>

git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@1118 80a73507-52ba-0310-bdc1-b9b4a1856fab